Moving sku block to blocks folder
