+++ packages/js/product-editor/src/blocks/inventory-sku/block.json
