+++ packages/js/product-editor/src/blocks/inventory-sku/edit.tsx
