+++ packages/js/product-editor/src/blocks/inventory-sku/index.ts
