+++ packages/js/product-editor/src/blocks/inventory-sku/style.scss
