+++ packages/js/product-editor/src/components/editor/init-blocks.ts
@@ -22,7 +22,7 @@ import { init as initPricing } from '../pricing-block';
 import { init as initCollapsible } from '../collapsible-block';
 import { init as initScheduleSale } from '../../blocks/schedule-sale';
 import { init as initTrackInventory } from '../../blocks/track-inventory';
-import { init as initSku } from '../inventory-sku-block';
+import { init as initSku } from '../../blocks/inventory-sku';
 
 export const initBlocks = () => {
 	const coreBlocks = __experimentalGetCoreBlocks();
